Created review card cpmonent
Created a review card component to display the reviews on the course. The card displays the name and the review. I added the component to the course details page.
We can notice some refactoring as well (prettier and deleting console.log())
We still need to get the user name. for now, we just display the user id